fix compiler warnings under 32 bit
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@17229 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/LogDlg.h
@@ -174,7 +174,7 @@ private:
 	void DiffSelectedRevWithPrevious();
 	void SetDlgTitle(bool bOffline);
 	CString GetAbsoluteUrlFromRelativeUrl(const CString& url);
-	void ToggleCheckbox(int item);
+	void ToggleCheckbox(size_t item);
 
 	inline int ShownCountWithStopped() const { return (int)m_logEntries.GetVisibleCount() + (m_bStrictStopped ? 1 : 0); }
